+++ CHANGELOG.md
@@ -21,6 +21,9 @@
 	- Added command to break on interrupt ("-e breakinterrupt on").
 - Settings:
 	- "esxdosRst" removed. See Reverse Engineering for replacement.
+- Switched to @vscode/debugadapter 1.57.0.
+- Use of debugadapter API:
+	- ASSERTIONs, WPMEMs and LOGPOINTs can now be easily enabled/disabled from the vscode BREAKPOINTS pane. The now superfluous DEBUG CONSOLE commands have been removed.
 
 
 # 2.7.4